+++ client/src/extension.ts
@@ -1,22 +1,13 @@
 import * as path from 'path';
-import { workspace as Workspace, ExtensionContext, workspace, commands, window as Window, 
-	Location, 
-} from 'vscode';
-
-import { 
-	LanguageClient, LanguageClientOptions, 
-	ServerOptions, TransportKind, RevealOutputChannelOn,
-	ExecuteCommandSignature, WorkspaceFolder
-} from 'vscode-languageclient/node';
-
-
+import * as vscode from 'vscode';
+import * as lc from 'vscode-languageclient/node';
 import * as ls from 'vscode-languageserver-protocol';
 
 import registerCommands from './common/commands';
 
 import CompositeDisposable from './common/compositeDisposable';
 
-let client: LanguageClient;
+let client: lc.LanguageClient;
 let disposables = new CompositeDisposable();
 
 interface ConfigurationSettings {
@@ -33,7 +24,7 @@ interface ConfigurationSettings {
 	}
 	lint?: Object;
 	keywords: Object[],
-	workspaceFolder?: WorkspaceFolder | undefined;
+	workspaceFolder?: lc.WorkspaceFolder | undefined;
 	callFunctions?: string[];
 }
 
@@ -42,7 +33,7 @@ interface CodeLensReferenceArgument {
 	locations: ls.Location[]
 }
 
-export function activate(context: ExtensionContext) {
+export function activate(context: vscode.ExtensionContext) {
 
 	// The server is implemented in node
 	let serverModule = context.asAbsolutePath(
@@ -50,39 +41,39 @@ export function activate(context: ExtensionContext) {
 	);
 
 	let debugOptions = { execArgv: ['--nolazy', '--inspect=6011'], cwd: process.cwd() };
-	let serverOptions: ServerOptions = {
-		run: { module: serverModule, transport: TransportKind.ipc, options: { cwd: process.cwd() } },
+	let serverOptions: lc.ServerOptions = {
+		run: { module: serverModule, transport: lc.TransportKind.ipc, options: { cwd: process.cwd() } },
 		debug: {
 			module: serverModule,
-			transport: TransportKind.ipc,
+			transport: lc.TransportKind.ipc,
 			options: debugOptions
 		}
 	};
 
-	let clientOptions: LanguageClientOptions = {
+	let clientOptions: lc.LanguageClientOptions = {
 		documentSelector: [{ language: 'macro', scheme: 'file' }],
-		initializationOptions: workspace.getConfiguration('macro'),
+		initializationOptions: vscode.workspace.getConfiguration('macro'),
 		synchronize: {
-			fileEvents: workspace.createFileSystemWatcher('**/*.{[sS][rR][cC],[dD][eE][fF],[lL][nN][kK]}')
+			fileEvents: vscode.workspace.createFileSystemWatcher('**/*.{[sS][rR][cC],[dD][eE][fF],[lL][nN][kK]}')
 		},		
 		diagnosticCollectionName: 'macro',
 		progressOnInitialization: true,
-		revealOutputChannelOn: RevealOutputChannelOn.Never,	
+		revealOutputChannelOn: lc.RevealOutputChannelOn.Never,	
 		middleware: {
-
-			executeCommand: async (command:string, args:any[], next:ExecuteCommandSignature) => {
+			
+			executeCommand: async (command:string, args:any[], next:lc.ExecuteCommandSignature) => {
 				if (command === 'macro.codelens.references') {
 					const arg:CodeLensReferenceArgument = args[0];
 
 
 					const position = client.protocol2CodeConverter.asPosition(arg.position);
-					const locations:Location[] = [];
+					const locations:vscode.Location[] = [];
 					for (const location of arg.locations){
 						locations.push(client.protocol2CodeConverter.asLocation(location));
 					}
 
-					if (Window.activeTextEditor) {
-						commands.executeCommand('editor.action.showReferences', Window.activeTextEditor.document.uri, position, locations);
+					if (vscode.window.activeTextEditor) {
+						vscode.commands.executeCommand('editor.action.showReferences', vscode.window.activeTextEditor.document.uri, position, locations);
 					}
 				}
 				else if (command === 'macro.action.refactorsequeces' || command === 'macro.action.addsequeces') {
@@ -90,29 +81,40 @@ export function activate(context: ExtensionContext) {
 						return Number.isInteger(Number(input)) ? null : 'Integer expected';
 					}
 
-					const config = workspace.getConfiguration('macro');
+					const config = vscode.workspace.getConfiguration('macro');
 					let start = undefined;
 					if (command === 'macro.action.refactorsequeces') {
-						start = await Window.showInputBox({
+						start = await vscode.window.showInputBox({
 							prompt: 'Start sequence number',
 							value: config.sequence.base,
 							validateInput: validate
 						});
 					}
 		
-					const increment = await Window.showInputBox({
+					const increment = await vscode.window.showInputBox({
 						prompt: 'Sequence number increment',
 						value: config.sequence.increment,
 						validateInput: validate
 					});
 
-					if (Window.activeTextEditor) {
+					if (vscode.window.activeTextEditor) {
 						if (command === 'macro.action.addsequeces' && increment) {
-							return next(command, [Window.activeTextEditor.document.uri.toString(), Window.activeTextEditor.selection.start, increment]);
+							return next(command, [vscode.window.activeTextEditor.document.uri.toString(), vscode.window.activeTextEditor.selection.start, increment]);
 						} 
 						else if (command === 'macro.action.refactorsequeces' && start && increment) {
-							return next(command, [Window.activeTextEditor.document.uri.toString(), Window.activeTextEditor.selection.start, start, increment]);
+							return next(command, [vscode.window.activeTextEditor.document.uri.toString(), vscode.window.activeTextEditor.selection.start, start, increment]);
 						}
+					}
+				}
+				else if (command === 'macro.action.validate') {
+					const workspaceUri = args[0];
+					if (workspaceUri) {
+						return next(command, [workspaceUri]);	
+					}
+					else {
+						pickFolder(workspace => {
+							return next(command, [workspace.uri.toString()]);	
+						});	
 					}
 				}
 			},
@@ -128,8 +130,8 @@ export function activate(context: ExtensionContext) {
 							continue;
 						}
 						const resource = client.protocol2CodeConverter.asUri(item.scopeUri);
-						const workspaceFolder = Workspace.getWorkspaceFolder(resource);
-						const config = workspace.getConfiguration('macro', workspaceFolder);
+						const workspaceFolder = vscode.workspace.getWorkspaceFolder(resource);
+						const config = vscode.workspace.getConfiguration('macro', workspaceFolder);
 						const settings: ConfigurationSettings = {
 							codelens: config.get('codelens'),
 							lint:config.get('lint', {}),
@@ -153,9 +155,8 @@ export function activate(context: ExtensionContext) {
 		} as any
 	};
 
-
 	// Create the language client and start the client.
-	client = new LanguageClient(
+	client = new lc.LanguageClient(
 		'macroLanguageServer',
 		'Macro Language Server',
 		serverOptions,
@@ -172,4 +173,25 @@ export function deactivate(): Thenable<void> | undefined {
 		return undefined;
 	}
 	return client.stop();
+}
+
+export function pickFolder(cb:(workspace: vscode.WorkspaceFolder) => void) {
+	const folders = vscode.workspace.workspaceFolders;
+	if (!folders) {
+		return;
+	}
+	
+	if (folders.length === 1) {
+		cb(folders[0]);
+		return;
+	}
+	vscode.window.showWorkspaceFolderPick({placeHolder:'', ignoreFocusOut:true}).then(selected => {
+		if (selected) {
+			cb(selected);
+		}
+	});
+	if (folders.length === 1) {
+		cb(folders[0]);
+		return;
+	}
 }