+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.3  (-, 2022)
+- Definition file dependency validation
+- Validation command
+
 ## 0.4.2  (July 6, 2022)
 - Fix a bug if an undefined symbol crashes the validator